Move tracker to hooks and fix cookie tests
