+++ src/App.tsx
@@ -1,15 +1,14 @@
 import './App.css';
 
-import { useContext } from 'react';
 import ShellPrompt from './components/ShellPrompt/ShellPrompt';
 import CookieNotice from './components/CookieNotice/CookieNotice';
-import { TRACKER_EVENTS, TrackerContext } from './Tracker';
+import { TRACKER_EVENTS, useTracker } from './hooks/useTracker';
 import { Notifications, useNotification } from './components/Notification/Notification';
 import { useEvent } from './hooks';
 
 function App() {
 
-    const tracker = useContext(TrackerContext);
+    const tracker = useTracker();
     const notifications = useNotification();
     useEvent('onAchievement', (achievement) => {
         notifications.add(`Achievement Unlocked: ${achievement.title}`, 5000);
