+++ src/hooks/useTracker.tsx
@@ -1,5 +1,5 @@
 import Plausible from 'plausible-tracker';
-import { createContext } from 'react';
+import { createContext, useContext } from 'react';
 
 export const TRACKER_EVENTS = Object.freeze({
     AchievementUnlocked: 'achievementUnlocked',
@@ -14,3 +14,5 @@ const tracker = Plausible({
 });
 
 export const TrackerContext = createContext(tracker);
+
+export const useTracker = () => useContext(TrackerContext);
